Split tests into VersionCatalogsSpec and ConventionCatalogsSpec
